Final commit

+++ src/components/VehicleList/VehicleList.js
@@ -3,6 +3,7 @@ import VehicleItem from '../VehicleItem';
 import Request from '../../utils';
 import { connect } from "react-redux";
 import {deleteVehicles} from '../../store/actions';
+import './VehicleList.scss';
 
 
 function mapDispatchToProps(dispatch){
@@ -19,22 +20,27 @@ function mapStateToProps(state) {
     
 class VehicleList extends Component{
     itemClick = (item) =>{
-       Request
-       .delete(`/vehicle/${item.id}`, {headers:{'Authorization' : "Bearer " + this.props.user.token}})
-       .then(result=>{
-          this.props.deleteVehicles(item.id);
-       })
-       .catch(error=>{
-           //exibir msg de erro
-       })
+        if(window.confirm("Deletar este veiculo?")){
+            Request
+            .delete(`/vehicle/${item.id}`, {headers:{'Authorization' : "Bearer " + this.props.user.token}})
+            .then(result=>{
+                this.props.deleteVehicles(item.id);
+            })
+            .catch(error=>{
+                alert('aconteceu algum erro, tente de novo em alguns instantes');
+            })
+        }
     }
     render(){
         let list = [...this.props.list];
         return(
-        <div>
+        <div class="panel">
+            <h2 className="panel-heading">Seus veiculos </h2>
+            
             {list.map(element => {
                 return (<VehicleItem key={element.id} plate={element.plate} onClick={this.itemClick.bind(null, element)}/>);
             })}
+            
         </div>
         )
     }
